+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-02-14
+## [Unreleased] - 2022-02-15
 
 ### Added
 
 ### Changed
 * Bumped dependency rho-mu to 1.2.0
 * Bumped dependency org.cicirello.core to 1.1.0
+* Migrated test cases to JUnit 5 (specifically JUnit Jupiter 5.8.2).
 
 ### Deprecated